+++ src/index.ts
@@ -1,3 +1,4 @@
+import enforcePlaceholders from './rules/enforce-placeholders.js'
 import noDynamicKey from './rules/no-dynamic-key.js'
 import preferStringArguments from './rules/prefer-string-arguments.js'
 import requireKey from './rules/require-key.js'
@@ -11,6 +12,7 @@ const plugin = {
     'require-key': requireKey,
     'prefer-string-arguments': preferStringArguments,
     'no-dynamic-key': noDynamicKey,
+    'enforce-placeholders': enforcePlaceholders,
   },
   configs: {} as Record<string, unknown>,
 }
@@ -23,6 +25,7 @@ plugin.configs.recommended = {
     'tolgee/require-key': 'error',
     'tolgee/prefer-string-arguments': 'error',
     'tolgee/no-dynamic-key': 'error',
+    'tolgee/enforce-placeholders': 'error',
   },
 }